feat: Complete model unit tests and getPromptAction unit tests
