+++ src/internal/ui/prompt/tokenize_test.go
@@ -1,0 +1,7 @@
+package prompt
+
+import "testing"
+
+func Test_tokenizePromptCommand(t *testing.T) {
+
+}
